feat(metrics): refactor RtmpEndpointMetrics to use RtmpRawMetrics and improve metrics handling
